Open YouTube page via WebKit rather than NSWorkspace.
No idea whether this is the _right_ thing to do but it certainly improves the behaviour in applications like NewsFire or BibDesk.

»Philosophically« the issue seems to be that going through NSWorkspace will send an AppleEvent to open the URL. If the application doesn't support the open URL AppleEvent, things start behaving strangely.

A further question would be whether it’d be preferable to open the YouTube.com page in a new window instead of opening it in the current one.