Add YouTube and Download links to film display.
Once a CtF view is clicked one loses the ability to visit the film’s YouTube page or to download its video file as was provided by the menu in the CtF view before. To restore that ability, links for that are provided beneath the film.

This should probably be tested on sites with exotic layouts to see how well it works. The idea is to not pick up extra spacing from the site’s CSS but respect the site’s font and link styling.