Add German Preference panel and other localised xibs.
This Preference Panel is redesigned. 
It leaves out the YouTube HD setting (why not just default to YES?).
It offers to quit the application after removing CtF. 

(Unfortunately it seems really hard to figure out the actual path of our bundle, so one could help the user more in case installation fails and/or locate it more reliably.)