Add an about box.
With the required menu item, text etc.

If rentzsch manages to construct a <a href>-able donation URL out of the PayPal stuff he built, there’s <a href> waiting to take it in Credits.html. Be sure to figure out how to show localised versions when doing so and give my hate to PayPal’s documentation while doing so.