Simplified the Sparkle loading code.
Assumes the presence of the required APIs since we check the framework version first. Removes calls to NSAssert since we aren't really set up to use it here (possible exceptions would be thrown into WebKit). Fails silently with a console message if any errors are encountered which results in business as usual minus Sparkle updates.