Apply Whitelisting to URL rather than host.
This lets you match variants of hostnames (google.com vs www.google.com) more easily and also matches things like google.com/analytics without whitelisting the whole google.com empire.

Do this in a way that is quite conservative and will not match the path only but always require the host name to be involved.

I hope there isn't a big reason for just applying it to the host lurking in the background.

http://rentzsch.lighthouseapp.com/projects/24342/tickets/361-google-analytics-aka-support-for-url-paths-in-approved-list