Use preview image in <video>, remove autobuffer.
Concerning the autobuffering:
It seems that if we want anything here it is to _not_ preload the film in case there are films on sites that are whitelisted. 

It reads like adding the autobuffer attribute with any content sets the element up to autobuffer.

Even when not adding the attribute, things are autobuffered, so this doesn't really help.