Improve ‘Larger Size’ download link text.
This hopefully clarifies the meaning of the link to avoid misunderstandings as in 

http://rentzsch.lighthouseapp.com/projects/24342/tickets/346-16b2-quicktime-larger-size-download

Also makes sure that the displayed links always match the actual situation. To achieve that all methods need to be passed the HD loading state rather than the URL to the video file.

Strings for the new links are added. fr localisation still missing.