Store videoHash explicitly and change its lookup.
The main aim of this was to make CtF work on YouTube user pages as well which need scraping despite being on YouTube.

To make the flow a bit clearer there is now an explicit videoHash variable and a single method, -setInfoFromFlash vars used to set the videoID and videoHash from Flash vars.